imp board with vec<vec<16>>

+++ src/core/svgbob/node.rs
@@ -21,7 +21,7 @@ pub enum GDirect {
 #[derive(Clone, Debug, Default, Eq, Hash)]
 pub struct GNode {
     // 节点排序用序号
-    pub idx: u8,
+    pub idx: u16,
     // 节点 id
     pub id: String,
     // 节点展示内容原始值
